refactor: consolidate field and Label compoennts
