+++ renderer/components/Form/index.js
@@ -7,10 +7,9 @@ export Checkbox from './Checkbox'
 export CryptoAmountInput from './CryptoAmountInput'
 export CurrencyFieldGroup from './CurrencyFieldGroup'
 export FiatAmountInput from './FiatAmountInput'
+export Field from './Field'
 export Form from './Form'
 export Input from './Input'
-export InputField from './InputField'
-export InputLabel from './InputLabel'
 export Label from './Label'
 export LightningInvoiceInput from './LightningInvoiceInput'
 export LndConnectionStringInput from './LndConnectionStringInput'
