+++ renderer/components/Home/WalletSettingsFormRemote.js
@@ -44,7 +44,6 @@ const WalletSettingsFormRemote = ({
             placeholder={intl.formatMessage({
               ...messages.connection_string,
             })}
-            rows={6}
             validateOnBlur
             validateOnChange
           />
